refactor(server): extract protocol-delivery state to core/protocol_delivery_state.py
Task D of docs/current/server-py-module-split.md. Pure extraction,
verbatim move verified byte-identical against git history (the
_PROTOCOL_* constants, _protocol_delivered_sessions/_protocol_call_counts
state, and the prune/mark/check helpers). Removed now-dead time and
OrderedDict imports from server.py. server.py temporarily imports the 6
names the tracker middleware still references; this import block is
removed in Task E when that middleware moves out. Functional check
exercises mark/check, TTL eviction, max-sessions cap, and call-count
prune from the new location, and confirms server.py shares the same
lock and dict objects.
